Skin Finished ver_0.2
Added Inline argument setting

+++ CSE169/include/Window.h
@@ -54,7 +54,7 @@ public:
 	static void resetCamera();
 
     // helper to change dof
-    static void changeDof(long value);
+    static void changeDof(float value);
     
     // helper to select joint
     static void selectJoint(int xyz);
